cli: Use the MetadataVersion from codegen
